requirements.txt
asgiref==3.5.0
Django==4.0.2
djangorestframework==3.13.1
iso3166==2.0.2
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
PyMySQL==1.0.2
